+++ milkman-auth/src/main/java/milkman/plugin/auth/oauth/DynamicOauth2Api.java
@@ -1,6 +1,7 @@
 package milkman.plugin.auth.oauth;
 
 import com.github.scribejava.core.builder.api.DefaultApi20;
+import com.github.scribejava.core.oauth2.clientauthentication.ClientAuthentication;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
@@ -9,4 +10,5 @@ import lombok.RequiredArgsConstructor;
 public class DynamicOauth2Api extends DefaultApi20 {
     private final String accessTokenEndpoint;
     private final String authorizationBaseUrl;
+    private final ClientAuthentication clientAuthentication;
 }